changed euler characteristic from size_t to int

+++ include/geometrycentral/surface/vertex_position_geometry.h
@@ -58,7 +58,7 @@ class VertexPositionGeometry : public EmbeddedGeometryInterface {
     double vertexMaxPrincipalCurvature(Vertex v) const;
 
     // CHANGED: for DDG
-    size_t eulerCharacteristic() const;
+    int eulerCharacteristic() const;
     double meanEdgeLength() const;
     double totalArea() const;
     double cotan(Halfedge he) const;
